+++ changelog.md
@@ -5,6 +5,7 @@
 * `NEW` hover function by keyword `function`
 * `NEW` code action: swap params
 * `CHG` unbind the relative path between binaries and scripts
+* `CHG` `LuaDoc` also catchs `--` (no need `---`)
 
 ## 1.5.0
 `2020-12-5`